Safari-611.1.10/LayoutTests/fast/text/font-display/parse-expected.txt

This test makes sure that the font-display descriptor in @font-face is appropriately parsed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sheet.rules[0].style.getPropertyValue('font-display') is ""
PASS sheet.rules[1].style.getPropertyValue('font-display') is ""
PASS sheet.rules[2].style.getPropertyValue('font-display') is ""
PASS sheet.rules[3].style.getPropertyValue('font-display') is "auto"
PASS sheet.rules[4].style.getPropertyValue('font-display') is "block"
PASS sheet.rules[5].style.getPropertyValue('font-display') is "swap"
PASS sheet.rules[6].style.getPropertyValue('font-display') is "fallback"
PASS sheet.rules[7].style.getPropertyValue('font-display') is "optional"
PASS sheet.rules[8].style.getPropertyValue('font-display') is ""
PASS window.getComputedStyle(document.getElementById('test9')).getPropertyValue('font-display') is ""
PASS successfullyParsed is true

TEST COMPLETE
